api+core: refactor to DDD; add domain/application/adapters/perf; remove engine
