+++ domain/bin/main/com/valr/domain/model/Level.kt
@@ -1,0 +1,9 @@
+package com.valr.domain.model
+
+import java.math.BigDecimal
+
+data class Level(
+    val price: BigDecimal,
+    val quantity: BigDecimal
+)
+
