+++ domain/bin/main/com/valr/domain/model/OrderBookSnapshot.kt
@@ -1,0 +1,8 @@
+package com.valr.domain.model
+
+data class OrderBookSnapshot(
+    val symbol: String,
+    val bids: List<Level> = emptyList(),
+    val asks: List<Level> = emptyList()
+)
+
