+++ domain/bin/main/com/valr/domain/model/TimeInForce.kt
@@ -1,0 +1,4 @@
+package com.valr.domain.model
+
+enum class TimeInForce { GTC, IOC, FOK }
+
