+++ domain/src/main/kotlin/com/valr/domain/model/Side.kt
@@ -1,0 +1,4 @@
+package com.valr.domain.model
+
+enum class Side { BUY, SELL }
+
