--- engine/bin/main/com/valr/engine/model/Level.kt
@@ -1,8 +1,0 @@
-package com.valr.engine.model
-
-import java.math.BigDecimal
-
-data class Level(
-    val price: BigDecimal,
-    val quantity: BigDecimal
-)
