--- engine/bin/main/com/valr/engine/model/Side.kt
@@ -1,6 +1,0 @@
-package com.valr.engine.model
-
-enum class Side {
-    BUY,
-    SELL
-}
