--- engine/bin/main/com/valr/engine/model/Trade.kt
@@ -1,12 +1,0 @@
-package com.valr.engine.model
-
-import java.math.BigDecimal
-
-data class Trade(
-    val id: String,
-    val price: BigDecimal,
-    val quantity: BigDecimal,
-    val takerOrderId: String,
-    val makerOrderId: String,
-    val timestamp: Long
-)
